+++ CHANGES.md
@@ -16,6 +16,7 @@
 - Add `for_each()`. Thanks [nathancahill]!
 - Add `for_each_right()`. Thanks [nathancahill]!
 - Add `assign()`. Thanks [nathancahill]!
+- Add `extend()`. Thanks [nathancahill]!
 - Add `remove()`.
 - Add `xor()`.
 - Add `sample()`.